(PC-17590)[PRO] chore: create component for OfferName
