+++ pro/src/components/pages/Offers/Offers/OfferItem/Cells/OfferNameCell/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OfferNameCell'
